+++ src/main/java/org/jacop/search/LargestMax.java
@@ -48,7 +48,7 @@ public class LargestMax<T extends IntVar> implements ComparatorVariable<T> {
     public LargestMax() {
     }
 
-    public int compare(float left, T var) {
+    public int compare(double left, T var) {
         int right = var.dom().max();
         if (left > right)
             return 1;
@@ -67,7 +67,7 @@ public class LargestMax<T extends IntVar> implements ComparatorVariable<T> {
         return 0;
     }
 
-    public float metric(T var) {
+    public double metric(T var) {
         return var.dom().max();
     }
