+++ src/main/java/org/jacop/search/PrioritySearch.java
@@ -581,7 +581,7 @@ public class PrioritySearch<T extends Var> extends DepthFirstSearch<T> {
 	    return n;
 
 	if (comparator != null) {
-	    float currentMeasure = comparator.metric(priority[current]);
+	    double currentMeasure = comparator.metric(priority[current]);
 	    for (int i = current+1; i < n; i++) {
 		if (comparator.compare(currentMeasure, priority[i]) < 0 && ! visited.get(i)) {
 		    current = i;
